requirements.txt
# Python
python>=3.7

# Django
Django>=3.0.5
Pillow
pytz
asgiref>=3.2.7
sqlparse>=0.2.2

# Heroku Deployment
gunicorn
django-heroku
whitenoise
psycopg2-binary
dj-database-url

# Rich Text Editor
django-ckeditor
